refactor: Dependency dropdown in separate component
